pyproject.toml
[tool.poetry]
name = "benchmark"
version = "0.1.0"
description = ""
authors = ["Ernesto Estevanell <ernestoestevanellv@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.8"
PyYAML = "^6.0"
yamlable = "^1.1.1"
black = "^22.8.0"
wget = "^3.2"
pandas = "^1.5.0"
numpy = "^1.23.3"
pathlib = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
